Move external imports to server.ts
